Refactor CustomTabs component for better modularity and maintainability
