+++ frontend/src/Component/CustomSnackbar.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { Snackbar, Alert } from "@mui/material";
+
+const CustomSnackbar = ({ open, message, severity, onClose }) => (
+  <Snackbar
+    open={open}
+    autoHideDuration={6000}
+    onClose={onClose}
+    anchorOrigin={{ vertical: "top", horizontal: "right" }}
+  >
+    <Alert onClose={onClose} severity={severity} sx={{ width: "100%" }}>
+      {message}
+    </Alert>
+  </Snackbar>
+);
+
+export default CustomSnackbar;
