+++ frontend/src/store/formStore.js
@@ -3,10 +3,9 @@ import { create } from 'zustand';
 const useFormStore = create((set) => ({
   formData: {},
   errors: {},
-  setFormData: (newFormData) => set((state) => ({ formData: { ...state.formData, ...newFormData } })),
+  setFormData: (data) => set((state) => ({ formData: { ...state.formData, ...data } })),
   clearFormData: () => set({ formData: {} }),
   setErrors: (errors) => set({ errors }),
-  clearErrors: () => set({ errors: {} }),
 }));
 
 export default useFormStore;
